requirements.txt
streamlit
tensorflow
numpy
Pillow
matplotlib
plotly
opencv-python-headless